관리자 페이지 레이아웃 수정
Update src/main/java/com/kh/bts/admin/controller/AdminCtrl.java
Update src/main/java/com/kh/bts/admin/model/dao/AdminDao.java
Update src/main/java/com/kh/bts/admin/model/service/AdminService.java
Update src/main/java/com/kh/bts/admin/model/service/AdminServiceImpl.java
Update src/main/resources/mappers/cashlog-mapper.xml
Update src/main/webapp/WEB-INF/views/admin/afterCommunityReport.jsp
Update src/main/webapp/WEB-INF/views/admin/afterReplyReport.jsp
Update src/main/webapp/WEB-INF/views/admin/cashEdit.jsp
Update src/main/webapp/WEB-INF/views/admin/cashLogList.jsp
Update src/main/webapp/WEB-INF/views/admin/communityReport.jsp
Update src/main/webapp/WEB-INF/views/admin/memberList.jsp
Update src/main/webapp/WEB-INF/views/admin/replyReport.jsp

+++ src/main/java/com/kh/bts/admin/controller/AdminCtrl.java
@@ -87,7 +87,7 @@ public class AdminCtrl {
 		return aService.totalCybcash();
 	}
 	@ModelAttribute("countTodayWon")
-	public int countTodayWon() {
+	public long countTodayWon() {
 		return aService.countTodayWon();
 	}
 	@ModelAttribute("countCreport")
